requirements.txt
# Core NLP libraries
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8
scikit-learn>=1.3.0
gensim>=4.3.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web interface
flask>=2.3.0
flask-cors>=4.0.0

# Text processing
beautifulsoup4>=4.12.0
requests>=2.31.0
newspaper3k>=0.2.8
clean-text>=0.6.0

# Sentiment analysis
textblob>=0.17.1
vaderSentiment>=3.3.2

# Statistical analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Report generation
jinja2>=3.1.0
weasyprint>=59.0
fpdf2>=2.7.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0